% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{get_census}
\usage{
get_census(geoid = NULL, year = 2010)
}
\arguments{
\item{geoid}{c(x,y,z) a vector of block group IDs to return}

\item{year}{The Census year. Currently only 2010 available.}
}
\description{
Load Census block group population data for Minnesota.
}
\examples{
# All block groups
bgs <- get_census()

# A single block group
bgs <- get_census(geoid = c(271090017024))
}
\keyword{block}
\keyword{blockgroup}
\keyword{census}
