% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closed_facilities.R
\name{get_closed_facilities}
\alias{get_closed_facilities}
\title{get_closed_facilities}
\usage{
get_closed_facilities(
  start_year = 2013,
  end_year = as.numeric(substring(Sys.Date(), 1, 4))
)
}
\arguments{
\item{start_year}{The start year to look for permit terminations. Default = 2010.}

\item{end_year}{The end year to look for terminations. Default = Sys.Date().}
}
\description{
Closed facilities from TEMPO
}
\examples{
# All receptors
closed_facilities <- get_closed_facilities()

#For non-default year
closed_facilities <- get_closed_facilities(year = 2025)
}
\keyword{Minnesota}
\keyword{permit}
\keyword{terminated}
