% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copc_groups.R
\name{get_copc_groups}
\alias{get_copc_groups}
\title{get_copc_groups}
\usage{
get_copc_groups(year = 2017)
}
\arguments{
\item{year}{The MNRISKS results year. Default value is 2017.}
}
\description{
MNRISKS pollutants / COPCs (chemicals of potential concern) groups
}
\examples{
# COPC groups for 2017
copc_groups <- get_copc_groups(year = 2017)

}
\keyword{copc}
\keyword{mnrisks}
\keyword{pollutant}
