% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copcs.R
\name{get_copcs}
\alias{get_copcs}
\title{get_copcs}
\usage{
get_copcs(year = 2017)
}
\arguments{
\item{year}{The MNRISKS results year. Default value is 2017.}
}
\description{
MNRISKS pollutants / COPCs (chemicals of potential concern) and parameters
}
\examples{
# COPC parameters for 2017
copcs <- get_copcs(year = 2017)

}
\keyword{copc}
\keyword{mnrisks}
\keyword{pollutant}
