% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facility_risks.R
\name{get_facility_risks}
\alias{get_facility_risks}
\title{get_facility_results}
\usage{
get_facility_risks(source_id = NULL, year = 2017)
}
\arguments{
\item{year}{The MNRISKS results year. Default value is 2017.}
}
\description{
MNRISKS modeled Resident Risks and air concentrations for facilities
}
\keyword{facility}
\keyword{mnrisks}
\keyword{risks}
