% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutants.R
\name{get_pollutants}
\alias{get_pollutants}
\title{get_pollutants}
\usage{
get_pollutants(year = 2014)
}
\arguments{
\item{year}{The MNRISKS results year. Default value is 2014.}
}
\description{
MNRISKS pollutants / COPCs (chemicals of potential concern) and parameters
}
\examples{
# Pollutant parameters for 2014
pollutants <- get_pollutants(year = 2014)

}
\keyword{copc}
\keyword{mnrisks}
\keyword{pollutant}
