% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receptor_bg_areas.R
\name{get_receptor_bg_areas}
\alias{get_receptor_bg_areas}
\title{get_receptor_bg_areas}
\usage{
get_receptor_bg_areas(bg_id = NULL, year = 2017)
}
\arguments{
\item{bg_id}{Block group IDs to include. Default includes all.}

\item{year}{The MNRISKS results year. Default is 2017.}
}
\description{
The modeled MNRISKS receptor area coverage for each block group
}
\examples{
# All receptors
rec_fractions <- get_receptor_bg_areas()

#For selected block groups
rec_fractions <- get_receptor_bg_areas(bg_id = c(271090017012))
}
\keyword{mnrisks}
\keyword{receptors}
