% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receptors.R
\name{get_receptors}
\alias{get_receptors}
\title{get_receptors}
\usage{
get_receptors(bg_id = NULL, year = 2017)
}
\arguments{
\item{bg_id}{Block group IDs to include. Default includes all.}

\item{year}{The MNRISKS results year. Default value is 2017.}
}
\description{
MNRISKS modeled receptors
}
\examples{
# All receptors
points <- get_receptors()

#For selected block groups
points <- get_receptors(bg_id = c(271090017012))
}
\keyword{mnrisks}
\keyword{receptors}
