% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_make_voronoi_polys.R, R/make_voronoi_polys.R
\name{make_voronoi_polys}
\alias{make_voronoi_polys}
\title{make_voronoi_polys}
\usage{
make_voronoi_polys()

make_voronoi_polys()
}
\arguments{
\item{year}{The MNRISKS results year. Default value is 2017.}
}
\description{
Create Minnesota voronoi grid from MNRISK receptors

Create Minnesota voronoi grid from receptors
}
\examples{
\dontrun{
# For all counties
make_voronoi_polys()
}

\dontrun{
# For all counties
make_voronoi_polys()
}

}
\keyword{grid}
\keyword{polygons}
\keyword{spatial}
\keyword{voronoi}
