% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bgs.R
\name{map_bgs}
\alias{map_bgs}
\title{map_bgs}
\usage{
map_bgs(
  data = NULL,
  result_col = "avg_risk",
  geoid_col = "geoid",
  colors = "viridis",
  reverse_colors = TRUE,
  signif_digits = 3,
  silent = FALSE
)
}
\arguments{
\item{data}{The dataframe or sf object containing the modeling results for mapping.}

\item{result_col}{Column name containing the modeling concentration/risk results.}

\item{geoid_col}{Block group GEOID colulmn name for mapping results.}

\item{colors}{Color palette to use for block group fill, ex. "viridis", "inferno", "Blues", "Greens"}

\item{reverse_colors}{Flip the color palette order}

\item{signif_digits}{Number of significant figures to show in map labels.}

\item{silent}{Return map as object. Default = FALSE}
}
\description{
Map the average block group mnrisks results with leaflet
}
\examples{
\dontrun{
# For all block groups
map_bgs(data = mnrisk_results,
        result_col = "avg_cancer_risk", 
        colors = "Blues")
}
}
\keyword{Map}
\keyword{average}
\keyword{blockgroup}
\keyword{leaflet}
\keyword{mnrisks}
