% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsk.R
\name{read_rsk}
\alias{read_rsk}
\title{read_rsk}
\usage{
read_rsk(
  file,
  scenarios = NULL,
  year = 2017,
  filter_word = NULL,
  aggregate_facs = FALSE
)
}
\arguments{
\item{file}{Path to .Rsk file or .zip file}

\item{scenarios}{Column names to assign to result columns. Default will assign generic numbers (e.g. "scenario_1")}

\item{year}{The MNRISKS results year. Default value is 2017.}

\item{filter_word}{Will only keep results with the entered text, such as "Benzene" or "A_Boilers". The default = NULL will keep all results.}

\item{aggregate_facs}{TRUE/FALSE. Will group_by pollutant and SCC to sum risks/concentrations. Default = FALSE}
}
\description{
Read MNRISKS modeled .Rsk result file
}
\examples{
\dontrun{
# Read in a single .rsk file
acrolein <- read_rsk("acrolein_results.rsk", year = 2017)

# Read in a single Zipped .rsk file
boilers <- read_rsk("area_results.zip", year = 2017, filter_word = "Boilers")
}
}
\keyword{mnrisks}
\keyword{results}
\keyword{rsk}
