% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closed_facilities.R
\name{get_closed_facilities}
\alias{get_closed_facilities}
\title{get_closed_facilities}
\usage{
get_closed_facilities(year = 2022)
}
\arguments{
\item{year}{The facility focus program year. Default value is 2022.}
}
\description{
Closed facilities from TEMPO
}
\examples{
# All receptors
closed_facilities <- get_closed_facilities()

#For non-default year
closed_facilities <- get_closed_facilities(year = 2025)
}
\keyword{Minnesota}
\keyword{permit}
\keyword{terminated}
