% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_voronoi_polys.R
\name{make_voronoi_polys}
\alias{make_voronoi_polys}
\title{make_voronoi_polys}
\usage{
make_voronoi_polys(year = 2017, census_year = 2020)
}
\arguments{
\item{year}{The MNRISKS results year. Default value is 2017.}
}
\description{
Create Minnesota voronoi grid from MNRISK receptors
}
\examples{
\dontrun{
# For all counties
make_voronoi_polys()
}

}
\keyword{grid}
\keyword{polygons}
\keyword{spatial}
\keyword{voronoi}
